treewide: src: Move AppConstants into consts
