+++ src/main/java/com/unconv/spring/config/logging/LoggingAspect.java
@@ -1,6 +1,6 @@
 package com.unconv.spring.config.logging;
 
-import com.unconv.spring.utils.AppConstants;
+import com.unconv.spring.consts.AppConstants;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.AfterThrowing;
