+++ src/main/java/com/unconv/spring/web/rest/BookingController.java
@@ -1,10 +1,10 @@
 package com.unconv.spring.web.rest;
 
+import com.unconv.spring.consts.AppConstants;
 import com.unconv.spring.domain.Booking;
 import com.unconv.spring.dto.BookingDTO;
 import com.unconv.spring.model.response.PagedResult;
 import com.unconv.spring.service.BookingService;
-import com.unconv.spring.utils.AppConstants;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
