+++ src/main/java/com/unconv/spring/web/rest/EnvironmentalReadingController.java
@@ -1,11 +1,11 @@
 package com.unconv.spring.web.rest;
 
+import com.unconv.spring.consts.AppConstants;
 import com.unconv.spring.domain.EnvironmentalReading;
 import com.unconv.spring.dto.EnvironmentalReadingDTO;
 import com.unconv.spring.model.response.MessageResponse;
 import com.unconv.spring.model.response.PagedResult;
 import com.unconv.spring.service.EnvironmentalReadingService;
-import com.unconv.spring.utils.AppConstants;
 import java.util.List;
 import java.util.UUID;
 import javax.validation.Valid;
