+++ src/main/java/com/unconv/spring/web/rest/HumidityThresholdController.java
@@ -1,10 +1,10 @@
 package com.unconv.spring.web.rest;
 
+import com.unconv.spring.consts.AppConstants;
 import com.unconv.spring.domain.HumidityThreshold;
 import com.unconv.spring.dto.HumidityThresholdDTO;
 import com.unconv.spring.model.response.PagedResult;
 import com.unconv.spring.service.HumidityThresholdService;
-import com.unconv.spring.utils.AppConstants;
 import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
